Refactor: 게시글 작성 컴포넌트의 이미지, 해쉬태그 분리
